extract `CommonRendererUtils` to top level
