+++ src/org/intellij/grammar/editor/BnfRuleLineMarkerProvider.java
@@ -18,7 +18,7 @@ import com.intellij.psi.util.PsiTreeUtil;
 import com.intellij.util.containers.ContainerUtil;
 import org.intellij.grammar.BnfIcons;
 import org.intellij.grammar.KnownAttribute;
-import org.intellij.grammar.generator.Renderer.CommonRendererUtils;
+import org.intellij.grammar.generator.CommonRendererUtils;
 import org.intellij.grammar.generator.RuleGraphHelper;
 import org.intellij.grammar.java.JavaHelper;
 import org.intellij.grammar.psi.BnfAttributes;
