+++ src/org/intellij/grammar/livePreview/LivePreviewParser.java
@@ -18,7 +18,7 @@ import it.unimi.dsi.fastutil.objects.Object2IntOpenHashMap;
 import org.intellij.grammar.KnownAttribute;
 import org.intellij.grammar.analysis.BnfFirstNextAnalyzer;
 import org.intellij.grammar.generator.*;
-import org.intellij.grammar.generator.Renderer.CommonRendererUtils;
+import org.intellij.grammar.generator.CommonRendererUtils;
 import org.intellij.grammar.generator.java.JavaRenderer;
 import org.intellij.grammar.psi.*;
 import org.intellij.grammar.psi.impl.GrammarUtil;
