+++ tests/org/intellij/grammar/BnfUtilTest.java
@@ -8,7 +8,7 @@ import com.intellij.testFramework.UsefulTestCase;
 import org.intellij.grammar.generator.Case;
 import org.intellij.grammar.generator.NameFormat;
 import org.intellij.grammar.generator.NameShortener;
-import org.intellij.grammar.generator.Renderer.CommonRendererUtils;
+import org.intellij.grammar.generator.CommonRendererUtils;
 import org.intellij.grammar.generator.java.JavaNameShortener;
 
 import java.util.ArrayList;
